+++ src/Oro/Bundle/WorkflowBundle/Tests/Unit/Model/WorkflowAssemblerTest.php
@@ -46,6 +46,10 @@ class WorkflowAssemblerTest extends \PHPUnit_Framework_TestCase
      */
     protected function createWorkflow()
     {
+        $doctrineHelper = $this->getMockBuilder('Oro\Bundle\EntityBundle\ORM\DoctrineHelper')
+            ->disableOriginalConstructor()
+            ->getMock();
+
         $aclManager = $this->getMockBuilder('Oro\Bundle\WorkflowBundle\Acl\AclManager')
             ->disableOriginalConstructor()
             ->getMock();
@@ -54,7 +58,7 @@ class WorkflowAssemblerTest extends \PHPUnit_Framework_TestCase
             ->disableOriginalConstructor()
             ->getMock();
 
-        return new Workflow($aclManager, $restrictionManager);
+        return new Workflow($doctrineHelper, $aclManager, $restrictionManager);
     }
 
     /**