+++ src/Oro/Bundle/WorkflowBundle/Tests/Unit/Serializer/Normalizer/WorkflowDataNormalizerTest.php
@@ -48,6 +48,9 @@ class WorkflowDataNormalizerTest extends \PHPUnit_Framework_TestCase
         $this->attributeManager = $this->getMockBuilder('Oro\Bundle\ActionBundle\Model\AttributeManager')
             ->disableOriginalConstructor()
             ->getMock();
+        $doctrineHelper = $this->getMockBuilder('Oro\Bundle\EntityBundle\ORM\DoctrineHelper')
+            ->disableOriginalConstructor()
+            ->getMock();
         $aclManager = $this->getMockBuilder('Oro\Bundle\WorkflowBundle\Acl\AclManager')
             ->disableOriginalConstructor()
             ->getMock();
@@ -57,7 +60,7 @@ class WorkflowDataNormalizerTest extends \PHPUnit_Framework_TestCase
         $this->workflow = $this->getMock(
             'Oro\Bundle\WorkflowBundle\Model\Workflow',
             array('getName'),
-            array($aclManager, $restrictionManager, null, $this->attributeManager, null)
+            array($doctrineHelper, $aclManager, $restrictionManager, null, $this->attributeManager, null)
         );
         $this->attribute = $this->getMock('Oro\Bundle\ActionBundle\Model\Attribute');
         $this->normalizer = new WorkflowDataNormalizer();