+++ src/Oro/Bundle/WorkflowBundle/CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG for BAP-10806
 =======================
-* Constructor for class `\Oro\Bundle\WorkflowBundle\Model\Workflow` was changed, first argument `\Oro\Bundle\WorkflowBundle\Model\EntityConnector` was removed.
-* Removed `\Oro\Bundle\WorkflowBundle\Model\EntityConnector` dependency form `\Oro\Bundle\WorkflowBundle\Model\Workflow` class
+* Updated class constructor `\Oro\Bundle\WorkflowBundle\Model\Workflow`, first argument is `\Oro\Bundle\EntityBundle\ORM\DoctrineHelper`.
+* Removed `\Oro\Bundle\WorkflowBundle\Model\EntityConnector` dependency form `\Oro\Bundle\WorkflowBundle\Model\Workflow` class.
 * Updated `\Oro\Bundle\WorkflowBundle\Model\WorkflowManager`, removed unnecessary calling of `resetWorkflowItem`.
 * Updated all Unit Tests to support new `\Oro\Bundle\WorkflowBundle\Model\Workflow`
 * Definition for `oro_workflow.prototype.workflow` was changed, removed `\Oro\Bundle\WorkflowBundle\Model\EntityConnector` dependency